Move types to be importable from "@opensystemslab/planx-core/types"
