+++ src/types/data.ts
@@ -1,0 +1,7 @@
+export type Value =
+  | string
+  | number
+  | boolean
+  | null
+  | Array<Value>
+  | { [key: string]: Value };
