+++ tsconfig.json
@@ -14,7 +14,6 @@
     "rootDir": "./src",
     "typeRoots": ["./node_modules/@types", "./@types"],
     "declaration": true,
-    "declarationDir": "./types",
     "outDir": "dist"
   },
   "include": [
